Fake Company + index.html
index.html + styles
---------------
just boiler plate
I want to design in penpot first before doing more

Fake Company
_________________
I decided the colorscheme, picked assets, and started the design a couple days ago.
I just imported the assets file from that day into this repo this morning.

I am not finished with the design file. Im currently making the design in penpot.
When i do finish i will upload the file to this repo.
The begining of the site i have finished, dont see the design changing much.

Was going for something very simple, this is the first full website i will design and finish
All me, except for assets

The Fake Company logo is AI generated.

The rest of the assets were found on upsplash

For right now the credit for these images is in the file  ASSETS/imageCredit

Eventually i will provide links to the images on the actual site

Im making the mobile layout first.
I might not even design the desktop version in penpot,
This site is very simple.
The plan is to make a 'donation' and a more in depth pricing chart
on seperate pages.
I have not designed those pages yet.
---------------
in other news.
I was gonna go to bed tonight, and then i decided not to,
and now im regretting it. if i sleep now ill be late for work i wanna
sleep.

But no,

daily commits he said.

at least i got a jump on todays commits.